+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Add `:exclude_presign` option for client stubbing to exclude non-sent request from presigned url at #api_requests (Github Issue #1866)
+
 3.24.1 (2018-08-13)
 ------------------