+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased Changes
 ------------------
 
-* Issue - Stop counting #presigned_url as #api_requests (Github Issue#1866)
+* Issue - Support opt-out counting #presigned_url as #api_requests (Github Issue#1866)
 
 1.17.0 (2018-07-11)
 ------------------